+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: dcsicsak <dcsicsak@student.42.fr>          +#+  +:+       +#+         #
+#    By: icseri <icseri@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/11/20 10:41:15 by icseri            #+#    #+#              #
-#    Updated: 2024/12/11 14:29:22 by dcsicsak         ###   ########.fr        #
+#    Updated: 2024/12/11 19:40:04 by icseri           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -27,10 +27,12 @@ COMMON_SRCS = parsing.c \
 		cast_rays_vertical.c \
 		minimap.c \
 		draw_utils.c \
+		movement_utils.c \
+		mlx_utils.c
 
-SRCS = main.c movement.c mlx_utils.c init.c ${COMMON_SRCS}
+SRCS = main.c movement.c init.c mlx.c ${COMMON_SRCS}
 
-BONUS_SRCS = main_bonus.c mlx_utils_bonus.c movement_bonus.c init_bonus.c ${COMMON_SRCS}
+BONUS_SRCS = main_bonus.c mlx_bonus.c movement_bonus.c init_bonus.c ${COMMON_SRCS}
 
 OBJS = ${SRCS:.c=.o}
