+++ main.c
@@ -6,7 +6,7 @@
 /*   By: icseri <icseri@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 10:43:22 by icseri            #+#    #+#             */
-/*   Updated: 2024/12/11 12:23:13 by dcsicsak         ###   ########.fr       */
+/*   Updated: 2024/12/11 14:48:51 by icseri           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,7 @@ int	main(int argc, char **argv)
 	init_data(&data);
 	data.map.is_bonus = false;
 	parsing(argc, argv, &data.map);
+	set_counts(&data);
 	allocate_textures(&data);
 	init_mlx(&data);
 	mlx_mouse_move(data.mlx, data.win, WIDTH / 2, HEIGHT / 2);
